+++ addon/mixins/ags-service-mixin.js
@@ -90,7 +90,7 @@ export default Mixin.create({
   * Returns stuff like version, services and folders
    */
   getServerInfo (url, options) {
-    debug(`WARNING: service-mixin.getServerInfo does not use ArcGIS Rest JS.`);
+    debug(`WARNING: service-mixin.getServerInfo does not use ArcGIS REST JS.`);
     const serviceUrl = parseServiceUrl(url);
     let service = url;
     if (serviceUrl) {
@@ -103,7 +103,7 @@ export default Mixin.create({
    * Get the authentication information from a server.
    */
   getAuthInfo (url, options) {
-    debug(`WARNING: service-mixin.getAuthInfo does not use ArcGIS Rest JS.`);
+    debug(`WARNING: service-mixin.getAuthInfo does not use ArcGIS REST JS.`);
     let server = url;
     const serverUrl = parseServerUrl(url);
     if (serverUrl) {