+++ addon/services/feature-service.js
@@ -7,7 +7,7 @@ import {
   updateFeatures,
   deleteFeatures,
   getAttachments
-} from "@esri/arcgis-rest-feature-service";
+} from "@esri/arcgis-rest-feature-layer";
 
 export default Service.extend(serviceMixin, layerMixin, {
 
@@ -26,15 +26,9 @@ export default Service.extend(serviceMixin, layerMixin, {
     const options = {
       url,
       featureId: id,
-      httpMethod: 'GET'
+      authentication: this.get('session.authMgr')
     }
-    return getAttachments(options)
-    .catch((err) => {
-      if (err.name === 'ArcGISAuthError') {
-        options.authentication = this.get('session.authMgr');
-        return getAttachments(options);
-      }
-    })
+    return getAttachments(options);
   },
 
   /**
@@ -66,7 +60,7 @@ export default Service.extend(serviceMixin, layerMixin, {
    * data.keywords
    */
   updateAttachment (url, data, token) {
-    debug(`WARNING: feature-service.updateAttachment does not use ArcGIS Rest JS.`);
+    debug(`WARNING: feature-service.updateAttachment does not use ArcGIS REST JS.`);
     // the feature id must already be embedded in the url
     url = url + '/updateAttachment?f=json';
     return this.attachmentsRequest(url, data, token);
@@ -101,7 +95,7 @@ export default Service.extend(serviceMixin, layerMixin, {
    * data.keywords
    */
   addAttachment (url, data, token) {
-    debug(`WARNING: feature-service.addAttachment does not use ArcGIS Rest JS.`);
+    debug(`WARNING: feature-service.addAttachment does not use ArcGIS REST JS.`);
     url = url + '/addAttachment?f=json';
     return this.attachmentsRequest(url, data, token);
   },
@@ -114,7 +108,7 @@ export default Service.extend(serviceMixin, layerMixin, {
       url,
       objectIds: [objectId],
       authentication: this.get('session.authMgr')
-    });
+    })
   },
 
   /**
@@ -134,7 +128,7 @@ export default Service.extend(serviceMixin, layerMixin, {
    * data.keywords
    */
   deleteAttachments (url, data, token) {
-    debug(`WARNING: feature-service.deleteAttachments does not use ArcGIS Rest JS.`);
+    debug(`WARNING: feature-service.deleteAttachments does not use ArcGIS REST JS.`);
     url = url + '/deleteAttachments?f=json';
     return this.attachmentsRequest(url, data, token);
   },
@@ -161,7 +155,7 @@ export default Service.extend(serviceMixin, layerMixin, {
   },
 
   attachmentsRequest (url, data, token) {
-    debug(`WARNING: feature-service.attachmentsRequest does not use ArcGIS Rest JS.`);
+    debug(`WARNING: feature-service.attachmentsRequest does not use ArcGIS REST JS.`);
     let options = {
       method: 'POST',
       mimeType: 'multipart/form-data',